Move and rename classes
