+++ packages/database/package.json
@@ -32,7 +32,9 @@
     "dotenv": "^8.2.0",
     "knex": "0.14.6",
     "lodash.clonedeep": "^4.5.0",
+    "lodash.groupby": "^4.6.0",
     "lodash.keyby": "^4.6.0",
+    "lodash.orderby": "^4.6.0",
     "moment": "^2.24.0",
     "os": "0.1.1",
     "pg": "8.5.1",
