+++ packages/meditrak-server/package.json
@@ -48,7 +48,6 @@
     "lodash.clonedeep": "^4.5.0",
     "lodash.groupby": "^4.6.0",
     "lodash.keyby": "^4.6.0",
-    "lodash.orderby": "^4.6.0",
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.25",
     "multer": "^1.4.1",
