--- packages/meditrak-server/src/database/SurveyResponseChangeHandler/index.js
@@ -1,1 +1,0 @@
-export { SurveyResponseChangeHandler } from './SurveyResponseChangeHandler';
