+++ packages/meditrak-server/src/database/index.js
@@ -7,4 +7,3 @@ export { SyncQueue } from './SyncQueue';
 export { createMeditrakSyncQueue } from './createMeditrakSyncQueue';
 export * from './utilities';
 export * from './constants';
-export * from './SurveyResponseChangeHandler';
