+++ packages/meditrak-server/src/database/models/Survey.js
@@ -5,22 +5,6 @@
 
 import { MaterializedViewLogDatabaseModel, DatabaseType, TYPES } from '@tupaia/database';
 
-export const PERIOD_GRANULARITIES = {
-  YEARLY: 'yearly',
-  QUARTERLY: 'quarterly',
-  MONTHLY: 'monthly',
-  WEEKLY: 'weekly',
-  DAILY: 'daily',
-};
-
-export const PERIOD_GRANULARITY_TO_MOMENT_UNIT = {
-  [PERIOD_GRANULARITIES.YEARLY]: 'year',
-  [PERIOD_GRANULARITIES.QUARTERLY]: 'quarter',
-  [PERIOD_GRANULARITIES.MONTHLY]: 'month',
-  [PERIOD_GRANULARITIES.WEEKLY]: 'isoWeek',
-  [PERIOD_GRANULARITIES.DAILY]: 'day',
-};
-
 class SurveyType extends DatabaseType {
   static databaseType = TYPES.SURVEY;
