+++ tupaia-packages.code-workspace
@@ -114,8 +114,6 @@
     "cSpell.words": [
       "Clazz",
       "Codeship",
-      "DHIS",
-      "Flattable",
       "Lesmis",
       "MediTrak",
       "Microservice",
@@ -129,12 +127,14 @@
       "cleanup",
       "debouncing",
       "deleters",
+      "dhis",
       "dialogs",
       "enroll",
       "enrollment",
       "enrollments",
       "eqeqeq",
       "eventdate",
+      "flattable",
       "geojson",
       "groupby",
       "heatmap",
@@ -146,6 +146,7 @@
       "middlewares",
       "migrator",
       "oucode",
+      "outdater",
       "parameterise",
       "postgresql",
       "preaggregate",
